chore(dashboards): adding openseach codex dashboards (#260)
## Changes

Adding opensearch dashboards

Fixes #(issue)

### Type of change

- [ ] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Documentation
- [ ] Refactor
- [x] Other: doc update

## Testing

<!-- How did you test this? What environment did you use? -->

## Proof it works

<!-- Screenshots, logs, test output, or other evidence showing your
changes work as expected -->

## Checklist

- [ ] I've tested my changes
- [ ] I've updated relevant documentation
- [ ] My code follows the project's style (run `cargo fmt` and `cargo
clippy`)
- [ ] All tests pass

## Notes

<!-- Anything else reviewers should know? Areas where you'd like
feedback? -->

---------

+++ mermin-ebpf/src/main.rs
@@ -276,7 +276,7 @@ fn try_flow_stats(ctx: &TcContext, direction: Direction) -> Result<i32, Error> {
             for i in 0..192 {
                 let offset = parsed_offset + i;
                 if let Ok(byte) = ctx.load::<u8>(offset) {
-                    flow_event.packet_data[i as usize] = byte;
+                    flow_event.packet_data[i] = byte;
                 } else {
                     break;
                 }
